Attempt to stop size_t conversion warning.
git-svn-id: http://paulgrenyer.net/svn/aeryn/trunk@69 48d1c4d4-5707-0410-b125-91f9528d8bec

+++ tests/test_runner_test.cpp
@@ -229,7 +229,7 @@ namespace Aeryn
 			testRunner.RunByName( "Test8", report );
 			testRunner.RunByName( "Test10", report );
 
-			IS_EQUAL( static_cast< unsigned int >( 5 ), static_cast< unsigned int >( report.TestCount() ) );
+			IS_EQUAL( 5, report.TestCount() );
 			IS_EQUAL( "Tests", report.TestSetName() );
 			Report::StoreType store = report.Store();
 			IS_EQUAL(  static_cast< unsigned int >( 5 ) , static_cast< unsigned int >( store.size() ) );
